requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
ipapi==1.0.4
python-dateutil==2.9.0.post0
requests==2.32.3
schedule==1.2.2
six==1.16.0
urllib3==2.2.3
